+++ next-docs/pages/components/accordion.tsx
@@ -10,7 +10,6 @@ import ContentOutsideSizes from '../../public/examples/accordion/ContentOutsideS
 import ControlOutside from '../../public/examples/accordion/ControlOutside';
 import Default from '../../public/examples/accordion/Default';
 import Disabled from '../../public/examples/accordion/Disabled';
-import Divider from '../../public/examples/accordion/Divider';
 import OpenDefault from '../../public/examples/accordion/OpenDefault';
 import Single from '../../public/examples/accordion/Single';
 import Sizes from '../../public/examples/accordion/Sizes';
@@ -89,11 +88,6 @@ const Example = () => {
         code={examples ? examples.ContentHeader : 'Loading'}
       />
       <Preview
-        title="With divider"
-        preview={<Divider />}
-        code={examples ? examples.Divider : 'Loading'}
-      />
-      <Preview
         title="Sizes"
         preview={<Sizes />}
         code={examples ? examples.Sizes : 'Loading'}
@@ -155,7 +149,7 @@ const Example = () => {
         ]}
       />
       <PropsTable
-        title="Accordion.Item"
+        title="Accordion.Item props"
         data={[
           {
             name: 'value',