+++ workspaces/core/src/switch/Switch.tsx
@@ -14,8 +14,8 @@ const Switch: FC<Props> = ({
   name,
   value,
   onChange,
-  onBgColor = 'bg-piccolo',
-  offBgColor = 'bg-beerus',
+  onBgColor, // deprecated
+  offBgColor, // deprecated
   onIcon,
   offIcon,
   className,
@@ -43,6 +43,7 @@ const Switch: FC<Props> = ({
         'block cursor-pointer rounded-full transition',
         getSwitchSize(size),
         enabled ? onBgColor : offBgColor,
+        !onBgColor && !offBgColor && 'bg-beerus moon-checked:bg-piccolo',
         disabled && 'opacity-30 cursor-not-allowed select-none',
         className
       )}