+++ next-docs/CHANGELOG.md
@@ -1,5 +1,83 @@
 # Change Log
 
+## 9.35.0
+
+### Minor Changes
+
+- Deploy 9.35.1
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.35.0
+  - @heathmont/moon-accordion@9.35.0
+  - @heathmont/moon-assets@9.35.0
+  - @heathmont/moon-charts@9.35.0
+  - @heathmont/moon-components@9.35.0
+  - @heathmont/moon-core@9.35.0
+  - @heathmont/moon-datepicker@9.35.0
+  - @heathmont/moon-draggabletable@9.35.0
+  - @heathmont/moon-icons@9.35.0
+  - @heathmont/moon-pagination@9.35.0
+  - @heathmont/moon-popover@9.35.0
+  - @heathmont/moon-select@9.35.0
+  - @heathmont/moon-sidebar@9.35.0
+  - @heathmont/moon-table@9.35.0
+  - @heathmont/moon-themes@9.35.0
+  - @heathmont/moon-utils@9.35.0
+  - @heathmont/moon-table-tw@9.35.0
+  - @heathmont/moon-themes-tw@9.35.0
+
+## 9.34.5
+
+### Patch Changes
+
+- Update GA
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@9.34.5
+  - @heathmont/moon-assets@9.34.5
+  - @heathmont/moon-charts@9.34.5
+  - @heathmont/moon-components@9.34.5
+  - @heathmont/moon-core@9.34.5
+  - @heathmont/moon-datepicker@9.34.5
+  - @heathmont/moon-draggabletable@9.34.5
+  - @heathmont/moon-icons@9.34.5
+  - @heathmont/moon-pagination@9.34.5
+  - @heathmont/moon-popover@9.34.5
+  - @heathmont/moon-select@9.34.5
+  - @heathmont/moon-sidebar@9.34.5
+  - @heathmont/moon-table@9.34.5
+  - @heathmont/moon-themes@9.34.5
+  - @heathmont/moon-utils@9.34.5
+  - @heathmont/moon-core-tw@9.34.5
+  - @heathmont/moon-table-tw@9.34.5
+  - @heathmont/moon-themes-tw@9.34.5
+
+## 9.34.4
+
+### Patch Changes
+
+- Add GA
+- Updated dependencies [undefined]
+  - @heathmont/moon-accordion@9.34.4
+  - @heathmont/moon-assets@9.34.4
+  - @heathmont/moon-charts@9.34.4
+  - @heathmont/moon-components@9.34.4
+  - @heathmont/moon-core@9.34.4
+  - @heathmont/moon-datepicker@9.34.4
+  - @heathmont/moon-draggabletable@9.34.4
+  - @heathmont/moon-icons@9.34.4
+  - @heathmont/moon-pagination@9.34.4
+  - @heathmont/moon-popover@9.34.4
+  - @heathmont/moon-select@9.34.4
+  - @heathmont/moon-sidebar@9.34.4
+  - @heathmont/moon-table@9.34.4
+  - @heathmont/moon-themes@9.34.4
+  - @heathmont/moon-utils@9.34.4
+  - @heathmont/moon-core-tw@9.34.4
+  - @heathmont/moon-table-tw@9.34.4
+  - @heathmont/moon-themes-tw@9.34.4
+
 ## 9.34.3
 
 ### Patch Changes