+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 9.35.0
+
+### Minor Changes
+
+- Deploy 9.35.1
+
+## 9.34.5
+
+### Patch Changes
+
+- Update GA
+
+## 9.34.4
+
+### Patch Changes
+
+- Add GA
+
 ## 9.34.3
 
 ### Patch Changes